Refactor system title rendering and separate it from card rendering
The refactor I did on accident and is completely unnecessary, and does change behaviour slightly. If there's no name and there is a colour, it adds `<code>` under a colourised `<span>` instead of colourising the `<code>` directly.

The separation from card rendering is actually somewhat useful. Separation from card rendering means they can both run in parallel, and will both complete as soon as they can instead of waiting the other one to finish, which means it takes slightly less time to see the first update.

Before, you would've seen an update once both were done. Now, you see an update when one of them finishes, and another update when the other one finishes (usually the title finishes before the cards do because it less requests, but not always).

This is only a difference in a few milliseconds though, so it's not that big of a deal.